feat: FilterableRestaurantTable 개발 및 테스트
