+++ src/components/FilterableRestaurantTable.tsx
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+import Restaurant from '../types/Restaurant';
+import RestaurantTable from './RestaurantTable';
+import SearchBar from './SearchBar';
+import extractCategories from '../utils/extractCategories';
+import filterRestaurants from '../utils/filterRestaurants';
+
+type FilterableRestaurantTableProps = {
+  restaurants: Restaurant[];
+}
+
+export default function FilterableRestaurantTable({
+  restaurants,
+}: FilterableRestaurantTableProps) {
+  const [filterText, setFilterText] = useState<string>('');
+  const [filterCategory, setFilterCategory] = useState<string>('전체');
+  const categories = extractCategories(restaurants);
+  const filteredRestaurants = filterRestaurants(restaurants, { filterText, filterCategory });
+
+  return (
+    <div>
+      <SearchBar
+        filterText={filterText}
+        categories={categories}
+        setFilterText={setFilterText}
+        setFilterCategory={setFilterCategory}
+      />
+      <RestaurantTable
+        restaurants={filteredRestaurants}
+      />
+    </div>
+  );
+}
